requirements.txt
discord.py[voice]
python-dotenv
Motor
jishaku
ping3
psutil
yt_dlp
requests